store tokens in TEXT or CLOBs
The current data format is translated into MySQL as varchar(255), which is not sufficient to store tokens. However, token size greatly depends on the systems, algorithms, and total information included in a token. It is thus recommended to store them in TEXT format: https://stackoverflow.com/a/5050559. Yet, this decision might decrease performance, since MySQL for instance does not keep TEXT fields ready in memory. In the future, we *might* have to reverse this decision.